Change the Litematcia schematic format slightly. Improve importing.
The TileEntity and Entity data are no longer double nested compound tags,
but instead the position data is now stored normally inside the data tag.

The schematic version was bumped to version 2, and there is support for
loading the previous version 1 format TEs and entities.

The Schematica schematic import was improved such that the
LitematicaSchematic will now grab the TE and entity data directly from
the SchematicaSchematic, instead of using the data written back from
the entities in the temporary world.