Add custom Chunk, ChunkProvider and World classes without light updates
There is still more to be done, but this version did already significantly
improve large schematic import times.

My test case was the SciCraft 720 wide Mesa Quarry, and the import time of
that went from ~8 minutes down to 4 seconds without the light updates
and other unnecessary updates in the temporary world.